[flow] NFC: Move code for cloning producers into RegionOpUtils
This commit prepares those functions to be shared among multiple
source files.
